third_party/WebKit/LayoutTests/fast/canvas/canvas-overloads-setStrokeColor-expected.txt

Test the behavior of CanvasRenderingContext2D.setStrokeColor() when called with different numbers of arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.setStrokeColor() threw exception TypeError: Failed to execute 'setStrokeColor' on 'CanvasRenderingContext2D': 1 argument required, but only 0 present..
PASS ctx.setStrokeColor('red') is undefined
PASS ctx.setStrokeColor(0) is undefined
PASS ctx.setStrokeColor(0, 0) is undefined
PASS ctx.setStrokeColor(0, 0, 0) threw exception TypeError: Failed to execute 'setStrokeColor' on 'CanvasRenderingContext2D': Valid arities are: [1, 2, 4, 5], but 3 arguments provided..
PASS ctx.setStrokeColor(0, 0, 0, 0) is undefined
PASS ctx.setStrokeColor(0, 0, 0, 0, 0) is undefined
PASS ctx.setStrokeColor(0, 0, 0, 0, 0, 0) threw exception TypeError: Failed to execute 'setStrokeColor' on 'CanvasRenderingContext2D': Valid arities are: [1, 2, 4, 5], but 6 arguments provided..
PASS successfullyParsed is true

TEST COMPLETE
